Stop unwind only if both program pointer and stack pointer stay same
We mix both dwarf and fp(aka 'frame-pointer') unwind and the rules are as following:
1.try unwind-table first
2.if failed to unwind by unwind-table, use frame-pointer for first 10 frames
As we could not get accurate stack pointer when using fp, we do not update sp(aka 'stack-pointer') which may lead to shorter stack frames.
Not all bins and libs inside system enable the fp(compile with -fno-omit-frame-pointer), the frames unwind by fp are not reliable which may
mislead fault location.

Signed-off-by: ericlee <liyu1@huawei.com>